Set up default collision audio
