+++ Assets/Scripts/ObjectSoundInfo.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class ObjectSoundInfo : MonoBehaviour
+{
+    public AudioClip collisionSound;
+}
