+++ Assets/Scripts/ObjectSoundInfo.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 412ea15b39df1d34885435d7c86ecd85
